compression for physical
